Refactor Let/Alias/Export
The let/alias/export/drop commands have been refactored into a builtins module.
